# Conflicts:
#	features/operational/contact_us/src/main/res/layout/layout_bottom_text_toolbar.xml

+++ features/merchant/product_add_edit/src/main/java/com/tokopedia/product/addedit/shipment/presentation/model/ShipmentInputModel.kt
@@ -19,5 +19,5 @@ data class ShipmentInputModel(
 
 @Parcelize
 data class CPLModel(
-    var shipmentServicesIds: ArrayList<Int>? = arrayListOf()
+    var shipmentServicesIds: ArrayList<Long>? = arrayListOf()
 ) : Parcelable
